chore: resolve PR comments by merging synsp & syndp into synw
