+++ internal/scan.go
@@ -56,8 +56,6 @@ import (
 	"github.com/Azure/azqr/internal/scanners/sigr"
 	"github.com/Azure/azqr/internal/scanners/sql"
 	"github.com/Azure/azqr/internal/scanners/st"
-	"github.com/Azure/azqr/internal/scanners/syndp"
-	"github.com/Azure/azqr/internal/scanners/synsp"
 	"github.com/Azure/azqr/internal/scanners/synw"
 	"github.com/Azure/azqr/internal/scanners/traf"
 	"github.com/Azure/azqr/internal/scanners/vm"
@@ -475,8 +473,6 @@ func GetScanners() []scanners.IAzureScanner {
 		&sigr.SignalRScanner{},
 		&sql.SQLScanner{},
 		&synw.SynapseWorkspaceScanner{},
-		&synsp.SynapseSparkPoolScanner{},
-		&syndp.SynapseSqlPoolScanner{},
 		&traf.TrafficManagerScanner{},
 		&st.StorageScanner{},
 		&vm.VirtualMachineScanner{},
