+++ internal/scanners/synw/rules_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/synapse/armsynapse"
 )
 
-func TestSynapseWorkspaceScanner_Rules(t *testing.T) {
+func TestSynapseWorkspaceScanner_WorkspaceRules(t *testing.T) {
 	type fields struct {
 		rule        string
 		target      interface{}
@@ -129,7 +129,7 @@ func TestSynapseWorkspaceScanner_Rules(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			s := &SynapseWorkspaceScanner{}
-			rules := s.GetRules()
+			rules := s.getWorkspaceRules()
 			b, w := rules[tt.fields.rule].Eval(tt.fields.target, tt.fields.scanContext)
 			got := want{
 				broken: b,
@@ -137,6 +137,122 @@ func TestSynapseWorkspaceScanner_Rules(t *testing.T) {
 			}
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("SynapseWorkspaceScanner Rule.Eval() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestSynapseWorkspaceScanner_SparkPoolRules(t *testing.T) {
+	type fields struct {
+		rule        string
+		target      interface{}
+		scanContext *scanners.ScanContext
+	}
+	type want struct {
+		broken bool
+		result string
+	}
+	tests := []struct {
+		name   string
+		fields fields
+		want   want
+	}{
+		{
+			name: "SynapseSparkPoolScanner SLA",
+			fields: fields{
+				rule:        "synsp-002",
+				target:      &armsynapse.Workspace{},
+				scanContext: &scanners.ScanContext{},
+			},
+			want: want{
+				broken: false,
+				result: "99.9%",
+			},
+		},
+		{
+			name: "SynapseSparkPoolScanner CAF",
+			fields: fields{
+				rule: "synsp-001",
+				target: &armsynapse.BigDataPoolResourceInfo{
+					Name: to.Ptr("synsp-test"),
+				},
+				scanContext: &scanners.ScanContext{},
+			},
+			want: want{
+				broken: false,
+				result: "",
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := &SynapseWorkspaceScanner{}
+			rules := s.getSparkPoolRules()
+			b, w := rules[tt.fields.rule].Eval(tt.fields.target, tt.fields.scanContext)
+			got := want{
+				broken: b,
+				result: w,
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("SynapseSparkPoolScanner Rule.Eval() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestSynapseWorkspaceScanner_SqlPoolRules(t *testing.T) {
+	type fields struct {
+		rule        string
+		target      interface{}
+		scanContext *scanners.ScanContext
+	}
+	type want struct {
+		broken bool
+		result string
+	}
+	tests := []struct {
+		name   string
+		fields fields
+		want   want
+	}{
+		{
+			name: "SynapseSqlPoolScanner SLA",
+			fields: fields{
+				rule:        "syndp-002",
+				target:      &armsynapse.Workspace{},
+				scanContext: &scanners.ScanContext{},
+			},
+			want: want{
+				broken: false,
+				result: "99.9%",
+			},
+		},
+		{
+			name: "SynapseSqlPoolScanner CAF",
+			fields: fields{
+				rule: "syndp-001",
+				target: &armsynapse.SQLPool{
+					Name: to.Ptr("syndp-test"),
+				},
+				scanContext: &scanners.ScanContext{},
+			},
+			want: want{
+				broken: false,
+				result: "",
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := &SynapseWorkspaceScanner{}
+			rules := s.getSqlPoolRules()
+			b, w := rules[tt.fields.rule].Eval(tt.fields.target, tt.fields.scanContext)
+			got := want{
+				broken: b,
+				result: w,
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("SynapseSqlPoolScanner Rule.Eval() = %v, want %v", got, tt.want)
 			}
 		})
 	}
